// src/sports/enums/priority-league.enum.ts
export enum PriorityLeague {
  PREMIER_LEAGUE = 'eng.1',
  CHAMPIONSHIP = 'eng.2',
  LA_LIGA = 'esp.1',
  SERIE_A = 'ita.1',
  BUNDESLIGA = 'ger.1',
  LIGUE_1 = 'fra.1',
  EREDIVISIE = 'ned.1',
  PRIMEIRA_LIGA = 'por.1',
  SCOTTISH_PREMIERSHIP = 'sco.1',
  BELGIAN_PRO_LEAGUE = 'bel.1',
  TURKISH_SUPER_LIG = 'tur.1',
  NPFL = 'nga.1',
  SOUTH_AFRICA_PREMIER_DIVISION = 'rsa.1',
  BRAZIL_SERIE_A = 'bra.1',
  ARGENTINA_PRIMERA = 'arg.1',
  LIGA_MX = 'mex.1',
  MLS = 'usa.1',
  SAUDI_PRO_LEAGUE = 'ksa.1',
}