// src/sports/enums/priority-club-competition.enum.ts
export enum PriorityClubCompetition {
  UEFA_CHAMPIONS_LEAGUE = 'uefa.champions',
  UEFA_EUROPA_LEAGUE = 'uefa.europa',
  UEFA_CONFERENCE_LEAGUE = 'uefa.europa.conf',
  CAF_CHAMPIONS_LEAGUE = 'caf.champions',
  CAF_CONFEDERATION_CUP = 'caf.confed',
  COPA_LIBERTADORES = 'conmebol.libertadores',
  COPA_SUDAMERICANA = 'conmebol.sudamericana',
  CONCACAF_CHAMPIONS_CUP = 'concacaf.champions',
}