// src/sports/enums/priority-international-competition.enum.ts
export enum PriorityInternationalCompetition {
  FIFA_WORLD_CUP = 'fifa.world',
  FIFA_WORLD_CUP_QUALIFIERS = 'fifa.worldq',
  AFCON = 'caf.nations',
  AFCON_QUALIFIERS = 'caf.nations_qual',
  UEFA_EURO = 'uefa.euro',
  UEFA_EURO_QUALIFIERS = 'uefa.euroq',
  UEFA_NATIONS_LEAGUE = 'uefa.nations',
  COPA_AMERICA = 'conmebol.america',
  CONCACAF_GOLD_CUP = 'concacaf.gold',
  CONCACAF_NATIONS_LEAGUE = 'concacaf.nations.league',
  AFC_ASIAN_CUP = 'afc.asian.cup',
}